Refactor Controllers classes
